Implement player spawning
Create a few transforms around the map. The PlayerSpawnController holds
these transforms as spawn points. When a player dies, the server starts
a timer and after it ends, it will tell the client to "respawn" itself
by enabling all of its scripts again.

It seems a bit easier to just disable/enable the player's scripts and
components instead of trying to explicitly delete the object and spawn
it again.