Fix delay when hitting and eliminating enemies.
Using the network variable for health introduces gross delay when
hitting an enemy multiple times in a second. If the players network
health was low enough where the next shot would kill them, it would
destroy the player in a responsive manner to the client that shot them.

However, hitting many times in a second would cause a delay as those
hits needed to update the network variable before the client could
safely assume they were dead.

Instead use a normal variable manually synced between clients for
health. Along with that, an internal health variable is used for what
the client expects that players health to be. So when hitting multiple
times, this internal health variable is decreased. If 0 or below, it is
safe to assume the player is dead and act accordingly. The server should
eventually broadcast out the damage to every client which will sync the
health variable together.

This also changes how the player is respawned. In order to get more
smooth respawning, every client set's the players transform to the spawn
point, then reenables the network transform after a delay. This is
because otherwise there is delay before clients see the player has
respawned, but there is no delay for the player being respawned. This of
course leads to an advantage