Handle generic and inherited weapons and projectiles
Introduce Projectile class which holds damage, speed, distance, hitscan
values and act accordingly. This allows new projectile prefabs to be
created easily.

Also move much of the internal Pistol logic to Weapon. ie. spawning
projectiles and muzzle flash effects are delegated to Weapon, not
Pistol. May not be needed, but a OnFire() callback is provided for
inherited Weapon classes.