Introduce shotgun weapon
Extend from the weapon class to create a new Shotgun weapon. It spreads
in a circular pattern utilized permanant bloom. Some other changes of
note is the ability to say number of projectiles per shot, and to
originate hitscan projectiles from the barrel rather than the fps
camera. This has no effect on where the projectile hits however