Recoil recovery - Part 1
Introduce recoil recovery, after firing, the crosshair/camera will
return to where the player starting originally shooting from.

However this needs a second part to reduce this behavior if the player
is accounting for recoil already. Which means if the player pulls down
on the mouse while fire to stay on target, when they stop firing, the
crosshair/camera should stay very close to the target. The recovery
would be most apparent if the recoil was not controlled