Recoil recovery and adjust for player compensation - Part 2
More simple then I thought. If the player now compensates for the
recoil. The recovery mechanic will only pull back down for the amount
the player did not adjust for. With the exception of a configurable
minimum it must pull down. (Although 0 might be fine for this)

Potential note: This works via event from the player's camera. When the
camera is moved it sends an event. The recoil system listens for this
and will use the values to adjust. But this assumes that the only other
time Rotate() is called on the player camera is for mouse input. So may
need to revisit if that changes