First and Third person meshes. Custom pistol model/anims
Use free asset for the player's third person model, and the first person
arms. Introduce a custom created Beretta M9 model. Along with a few
starting animations for the fp model and the pistol (Idle, Fire, Reload,
and ReloadEmpty).

Third person meshes will be hidden for the Owner player and visble for
everyone else, and the first person meshes are the opposite