Show game is scoreboard and timer. Support timer victory and draw
The scoreboard now will show the game mode name in the corner. As well
as the current timer value for the time remaining in the game. When the
game ends, the gamemode implements a new function to check for the
highest scoring winner based on their gamemode rules. If draws are
possible than the text "Draw" will display instead of "Player Won"